+++ Nm-otool/Libft/ft_memcpy.c
@@ -3,18 +3,18 @@
 /*                                                        :::      ::::::::   */
 /*   ft_memcpy.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: plogan <marvin@42.fr>                      +#+  +:+       +#+        */
+/*   By: plogan <plogan@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/04/10 17:18:39 by plogan            #+#    #+#             */
-/*   Updated: 2017/04/11 17:45:36 by plogan           ###   ########.fr       */
+/*   Updated: 2019/08/28 13:13:45 by peterlogan       ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	*ft_memcpy(void *dst, const void *src, size_t n)
+void	*ft_memcpy(void *dst, const void *src, uint64_t n)
 {
-	size_t			i;
+	uint64_t			i;
 	unsigned char	*temp_dest;
 	unsigned char	*temp_src;
