+++ Nm-otool/Libft/libft.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   libft.h                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: plogan <marvin@42.fr>                      +#+  +:+       +#+        */
+/*   By: plogan <plogan@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/04/10 16:40:10 by plogan            #+#    #+#             */
-/*   Updated: 2017/06/07 18:40:11 by plogan           ###   ########.fr       */
+/*   Updated: 2019/08/28 13:13:55 by peterlogan       ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -42,7 +42,7 @@ typedef struct		s_list
 */
 void				*ft_memset(void *str, int c, size_t n);
 void				ft_bzero(void *s, size_t n);
-void				*ft_memcpy(void *dst, const void *src, size_t n);
+void				*ft_memcpy(void *dst, const void *src, uint64_t n);
 void				*ft_memccpy(void *dst, const void *src, int c,
 		size_t n);
 void				*ft_memmove(void *dst, const void *src, size_t len);
@@ -95,7 +95,7 @@ void				ft_putchar_fd(char c, int fd);
 void				ft_putstr_fd(char const *s, int fd);
 void				ft_putendl_fd(char const *s, int fd);
 void				ft_putnbr_fd(int n, int fd);
-t_list				*ft_lstnew(void const *content, size_t content_size);
+t_list				*ft_lstnew(void const *content, uint64_t content_size);
 void				ft_lstdelone(t_list **alst, void (*del)(void *, size_t));
 void				ft_lstdel(t_list **alst, void (*del)(void *, size_t));
 void				ft_lstadd(t_list **alst, t_list *new);
